% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqID_HRegSplines.R
\name{FreqID_HRegSplines}
\alias{FreqID_HRegSplines}
\title{The function to fit flexible parametric B-spline models for the frequentist anlaysis of semi-competing risks data
arising from cohort or nested case-control studies.}
\usage{
FreqID_HRegSplines(Formula, data, wts, nKnots, startvals,
  penalty = FALSE, Kappa = c(0, 0, 0), na.action = "na.fail",
  subset = NULL)
}
\arguments{
\item{Formula}{a \code{Formula} object of the form \eqn{y_1+\delta_1 | y_2+\delta_2 \sim x_1 | x_2 | x_3}. 
See the \code{SemiCompRisks} package for more details.}

\item{data}{a data.frame in which to interpret the variables named in \code{Formula}.}

\item{wts}{a vector of weights for each individual in \code{data}. Must be of length \code{nrow(data)}.}

\item{nKnots}{a vector of length 3 giving the number of knots used in the B-spline specification of each baseline hazard. The
length of \eqn{\phi_i, i\in\{1,2,3\}} in \code{startvals} is equal to \code{nKnots[i]+1}, which includes an intercept.}

\item{startvals}{a vector of starting values in the order (\eqn{\log\theta, \beta_1, \beta_2, \beta_3,
\phi_1, \phi_2, \phi_3}), where \eqn{\beta} represent regression coefficients and \eqn{\phi} represent the B-spline coefficients.}

\item{penalty}{an indicator of whether penalized maximum likelihood estimates are to be used. Defaults to \code{FALSE}.}

\item{Kappa}{a vector of length 3 providing values of smoothing parameters. Defaults to \code{c(0,0,0)}.}

\item{na.action}{how NAs are treated. See \code{model.frame}.}

\item{subset}{a specification of the rows to be used: defaults to all rows. See \code{model.frame}.}
}
\value{
\code{FreqID_HRegSplines} returns a list with two elements. The first element provides maximum likelihood
or maximum penalized likelihood estimates of the parameters, in the order 
(\eqn{\log\theta, \beta_1, \beta_2, \beta_3, \phi_1, \phi_2, \phi_3}). The second element is a vector of sandwich 
estimates of the standard error, in the same order.
}
\description{
Independent semi-competing risks data can be analyzed using hierarchical models. As currently implemented, the semi-Markov 
assumption can be adopted for the conditional hazard function for time to the terminal event given time to non-terminal event.
This function is a flexible parametric analog of \code{FreqID_HReg} in the R package \code{SemiCompRisks} that
accommodates user-specified weights, permitting the analysis of semi-competing risks data arising from a nested case-
control study.
}
\examples{
data(IDSplineData)
form <- Formula(y1 + delta1 | y2 + delta2 ~ cov1 + cov2 | cov1 + cov2 | cov1 + cov2)
### starting values for the baseline hazard components (phi)
### based on a spline approximation to a BayesID fit in SemiCompRisks. other strategies possible.
### each is length nKnots[i]+1
phi1.start <- c(-9.488466, -7.372098, -6.656212, -5.972695, -5.622274, -5.417296, -5.340710)
phi2.start <- c(-8.641889, -6.842977, -6.234473, -5.653484, -5.355626, -5.181395, -5.116297)
phi3.start <- c(-4.563376, -3.004653, -3.035086, -2.379959, -2.498012, -2.083923, -2.279812)
### starting values
startvals <- c(log(3), c(1,1), c(-1,-1), c(-0.5,-0.5), phi1.start, phi2.start, phi3.start) 
### 
fit1<- FreqID_HRegSplines(Formula=form, data=IDSplineData, wts=rep(1,nrow(IDSplineData)), 
nKnots=rep(6,3), startvals=startvals, penalty=FALSE)

}
